feat: migrate learn page to ssr
